+++ ParallaxGenLib/include/NIFUtil.hpp
@@ -90,6 +90,7 @@ auto configureShaderFlag(nifly::BSShaderProperty *NIFShaderBSLSP, const nifly::S
 auto setTextureSlot(nifly::NifFile *NIF, nifly::NiShape *NIFShape, const TextureSlots &Slot, const std::wstring &TexturePath, bool &Changed) -> void;
 auto setTextureSlot(nifly::NifFile *NIF, nifly::NiShape *NIFShape, const TextureSlots &Slot,
                     const std::string &TexturePath, bool &Changed) -> void;
+auto getTextureSlot(nifly::NifFile *NIF, nifly::NiShape *NIFShape, const TextureSlots &Slot) -> std::string;
 auto getTexBase(const std::filesystem::path &TexPath) -> std::wstring;
 auto getTexMatch(const std::wstring &Base,
                  const std::map<std::wstring, PGTexture> &SearchMap) -> PGTexture;