+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 - Added --high-mem option for paster processing in exchange for high memory usage
 - Added wide string support in filenames
 - Fixed issue of blank DDS files being checked for aspect ratio
+- CM will be rejected on shapes with textures in slots 3, 7, or 8 now
+- Added icon to parallaxgen.exe
 
 ## [0.5.7] - 2025-09-05